A significant refactoring of how links and views are generated in the context of other applications.
You can now do:

request.parent.link(..)

and

request.child(app, **variables).link(...)

Also reworked the way mounting context works, so that returning None
will now trigger a 404; the mounting does not apply to that variable.
That brings it more in line with the way paths work. In addition, make
sure that this is checked when .child() is used, so that an unknown
child will use the fallback link generator.