The Mount class is now gone. Instead we reuse the App subclasses for mounting. This simplifies the mount code considerably.
There are a few breaking changes introduced with this code, however: the
mount directive now takes (and sometimes rqeuires) a variables argument,
and request.child() takes the arguments to the App class (i.e. what's
in the dictionary returned from the mount function) instead of the path
variables as it did before.