function directive is really about functions.
The function directive shouldn't accept app as an optional first
argument; it's really for registering functions. Instead use the
method directive, which is now tested, which always takes app
as an argument. Refactor verify_identity directive so it uses
Reg registration direction instead of generating a FunctionDirective,
as it *does* expect an optional app first argument.