Make the traject tests work again.
This includes fixing a bug concerning assumptions about parameter
generation that were incorrect. Previously I thought I could store
the parameter names expected in the node, but this won't work if
the node is not an end-node and contains parameters. So I cannot avoid
something something like mapply, at least for now.