Fix traject.
This fixes the absorb behavior. This is also a much better fix for the
nested path bug than the one before. The insight is that if consume
returns None, for whatever reason, we're done immediately (and this is
a 404 error).

If consumption ends at an App instance then we need to go to the root.
We can do this unconditionally as we know there is nothing left to
consume.